requirements.txt
Flask
gunicorn
pandas
sentence-transformers
pinecone-client
requests